% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.R
\name{angle}
\alias{angle}
\title{Calculate the angle between two vectors}
\usage{
angle(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric value.
}
\description{
Calculates the angle in radians between two vectors.
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
