% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairangle.R
\name{pairangle}
\alias{pairangle}
\title{Pairwise angles}
\usage{
pairangle(mat)
}
\arguments{
\item{mat}{A numeric matrix with each row representing a vector.}
}
\value{
A square matrix whose \code{[i,j]} entry is the angle between the two vectors
represented in row \code{i} and \code{j} of \code{mat}.
}
\description{
Computes the matrix of angles between all pairs of vectors.
}
\details{
The (i, j)-entry is the angle between the vectors represented by the i'th and j'th row of the input matrix
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
