\name{print.gene_ontology}
\alias{print.gene_ontology}
\docType{package}
\title{
  Show gene ontology tree in dot format
}
\description{
  Show gene ontology tree in dot format
}
\usage{
\method{print}{gene_ontology}(x, \dots)
}
\arguments{
  \item{x}{
    Tree object.
  }
  \item{...}{
    Ignored.
  }
}
\value{
  Produces a representation of the gene ontology tree in dot format.
}
\seealso{
  \code{\link[treeFun]{tree2dot}}
  \code{\link[treeFun]{print.tree}}
}
