% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichdiffincyl}
\alias{whichdiffincyl}
\title{Is the difference of two points contained in a cylinder?}
\usage{
whichdiffincyl(mat, dir, r_sq, t)
}
\arguments{
\item{mat}{A matrix in which the rows represent a point in some three dimensional space.}

\item{dir}{A vector of the same dimension as the points.}

\item{r_sq}{A vector or single value of squared radii of a cylinder.}

\item{t}{A vector or single value of heights of a cylinder.}
}
\value{
A \code{\link{list}} of \code{\link{data.frame}s}
where each list element is a unique combination of radius and height of the cylinder.
Each \code{\link{data.frame}} have two columns: \code{index1} and \code{index2}, s.t.
\code{mat[index2, ] - mat[index1, ]} is contained in the specified cylinder.
}
\description{
The function finds out if the vector between two point \code{x_1} and \code{x_2}
is contained in a specified cylinder.
Moreover, is \code{x_2} contained in some cylinder extending from \code{x_1}.
}
\details{
This function also works for two dimensions. In which case \code{r_sq} is half the width squared
}
